+++ src/discord.py
@@ -1,5 +1,5 @@
 ##
-# Part of `SmartNodeMonitorBot`
+# Part of `StashNodeMonitorBot`
 #
 # Copyright 2018 dustinface
 #
@@ -38,13 +38,13 @@ from src import faq
 from src.commandhandler import node
 from src.commandhandler import user
 from src.commandhandler import common
-from src.smartexplorer import WebExplorer
+from src.stashexplorer import WebExplorer
 
-from smartcash.rewardlist import SNReward
+from stashcash.rewardlist import SNReward
 
 logger = logging.getLogger("bot")
 
-class SmartNodeBotDiscord(object):
+class StashNodeBotDiscord(object):
 
     def __init__(self, botToken, admins, password, db, nodeList, rewardList):
 
@@ -161,7 +161,7 @@ class SmartNodeBotDiscord(object):
 
             with self.nodeList as nodeList:
 
-                await self.client.change_presence(game=discord.Game(name='our {} SmartNodes'.format(nodeList.count()), type=3))
+                await self.client.change_presence(game=discord.Game(name='our {} StashNodes'.format(nodeList.count()), type=3))
 
                 # # Update the sources where the blocks are assigned to the nodelist
                 # for node in nodeList.nodes.values():
@@ -237,7 +237,7 @@ class SmartNodeBotDiscord(object):
             args = parts[2:]
         # If there are multiple mentions send each one (excluded the bot itself)
         # the help message.
-        # Like: hey @dustinface and @whoever check out the @SmartNodeMonitorBot
+        # Like: hey @dustinface and @whoever check out the @StashNodeMonitorBot
         # The above would send @dustinface and @whoever the help message of the bot.
         elif len(message.mentions) > 1 and self.client.user in message.mentions:
 
@@ -498,23 +498,23 @@ class SmartNodeBotDiscord(object):
 
 
     ######
-    # Callback which get called when there is a new releases in the smartcash repo.
+    # Callback which get called when there is a new releases in the stashcash repo.
     #
-    # Called by: Nothing yet, SmartGitHubUpdates later.
+    # Called by: Nothing yet, StashGitHubUpdates later.
     #
     ######
     def updateCheckCallback(self, tag):
 
         for user in self.database.getUsers():
             self.sendMessage(user['id'], ("*Node update available*\n\n"
-                                         "https://github.com/SmartCash/smartcash/releases/tag/{}").format(tag))
+                                         "https://github.com/StashCash/stashcash/releases/tag/{}").format(tag))
 
 
     ######
     # Callback for evaluating if someone in the database had an upcomming event
     # and send messages to all chats with activated notifications
     #
-    # Called by: SmartNodeList
+    # Called by: StashNodeList
     #
     ######
     def nodeUpdateCB(self, update, n):
@@ -534,7 +534,7 @@ class SmartNodeBotDiscord(object):
     # Callback for evaluating if someone in the database has won the reward
     # and send messages to all chats with activated notifications
     #
-    # Called by: SNRewardList from python-smartcash
+    # Called by: SNRewardList from python-stashcash
     #
     ######
     def rewardCB(self, reward, distance):
@@ -560,7 +560,7 @@ class SmartNodeBotDiscord(object):
     ######
     # Callback for SNRewardList errors
     #
-    # Called by: SNRewardList from python-smartcash
+    # Called by: SNRewardList from python-stashcash
     #
     ######
     def rewardListErrorCB(self, error):
@@ -570,13 +570,13 @@ class SmartNodeBotDiscord(object):
     # Callback for evaluating if someone has enabled network notifications
     # and send messages to all relevant chats
     #
-    # Called by: SmartNodeList
+    # Called by: StashNodeList
     #
     ######
     def networkCB(self, collaterals, added):
 
         nodeCount = self.nodeList.count()
-        asyncio.run_coroutine_threadsafe(self.client.change_presence(game=discord.Game(name='our {} SmartNodes'.format(nodeCount), type=3)), loop=self.client.loop)
+        asyncio.run_coroutine_threadsafe(self.client.change_presence(game=discord.Game(name='our {} StashNodes'.format(nodeCount), type=3)), loop=self.client.loop)
 
         response = common.networkUpdate(self, collaterals, added)
 
@@ -609,10 +609,10 @@ class SmartNodeBotDiscord(object):
             self.database.deleteNodesWithId(collateral)
 
     ######
-    # Callback which gets called from the SmartNodeList when a balance request triggered by any user
+    # Callback which gets called from the StashNodeList when a balance request triggered by any user
     # is done. It sends the result to the related user.
     #
-    # Called by: SmartExplorer
+    # Called by: StashExplorer
     #
     ######
     def balancesCB(self, check, results):
@@ -643,7 +643,7 @@ class SmartNodeBotDiscord(object):
     ######
     # Push the message to the admin
     #
-    # Called by: SmartNodeList
+    # Called by: StashNodeList
     #
     ######
     def adminCB(self, message):