+++ src/faq.py
@@ -1,5 +1,5 @@
 ##
-# Part of `SmartNodeMonitorBot`
+# Part of `StashNodeMonitorBot`
 #
 # Copyright 2018 dustinface
 #
@@ -88,7 +88,7 @@ def position(bot):
         top10Time = util.secondsToText(top10Seconds)
 
         return (
-        "We have currenty <b>{}<b> qualified SmartNodes. All those are in a "
+        "We have currenty <b>{}<b> qualified StashNodes. All those are in a "
         "virtual payout queue. Every other mined block (2*55seconds) one "
         "of the nodes in the top 10% of that queue gets picked perchance and"
         " receives a payout. After that it moves back to the end of the queue.\n\n"
@@ -102,7 +102,7 @@ def position(bot):
         "bit in the queue from time to time. This happens each time a node that"
         " has received its last payment longer ago as yours becomes eligible"
         " for payouts (it jumpes into the queue and receives a position)."
-        " At this moment we have <b>{}<b> SmartNodes. <b>{}<b> of them are"
+        " At this moment we have <b>{}<b> StashNodes. <b>{}<b> of them are"
         " <b>not<b> qualified for payouts. Each time one of the unqualified"
         " nodes becomes eligible due to a full match of the requirements"
         " it is very likely that it will jump ahead of yours.\n\n"
@@ -123,7 +123,7 @@ def collateral(bot):
 
         return (
         "A too new collateral means that your nodes collateral transaction (the 100k one) does not"
-        " have the minimum required number of confirmations in the SmartCash blockchain."
+        " have the minimum required number of confirmations in the StashCash blockchain."
         "This number of confirmations is currently <b>{}<b>.\nYour collateral gets 1 confirmation with each"
         " new mined block.\n\nMeans right now you need to wait {} x 55 seconds => <b>~{}<b> until"
         " your collateral transaction matches the requirement.\n\n"
@@ -181,19 +181,19 @@ def rewards(bot):
         perMonth = round((30.5 / payoutDays) * currentReward,1)
 
         return (
-        "The SmartNode rewards are calculated by the following formula\n\n"
+        "The StashNode rewards are calculated by the following formula\n\n"
         "```reward = 5000 x 143500 / blockHeight * 0.1```\n\n"
         "At this moment our blockchain is at the height <b>{}<b> that means"
         "\n\n```5000 x 143500 / {} * 0.1 => {} SMART per block```\n\n"
         "Each block with an <b>even<b> blockheight one of the the nodes receive this reward for 2 blocks. With the current "
         "estimated payout interval of <b>{}<b> you can expect roughly"
-        " <b>{:,} SMART<b> per month per SmartNode. This can vary a bit upwards and downwards though.\n\n"
-        "Due to the constant increase of the <c>blockHeight<c> of the SmartCash blockchain"
+        " <b>{:,} SMART<b> per month per StashNode. This can vary a bit upwards and downwards though.\n\n"
+        "Due to the constant increase of the <c>blockHeight<c> of the StashCash blockchain"
         " the rewards will decrease a little bit every 55 seconds."
         " Also the increase of the number of qualified nodes will increase the payout interval."
         " As result your monthly payout will slightly decrease over the time.\n\n"
         "You can look at the chart in the link below to see the reward decrease "
-        "for the first 4 years after the SmartNode launch.\n\n"
+        "for the first 4 years after the StashNode launch.\n\n"
         ).format(lastBlock, lastBlock, currentReward, interval, perMonth)\
         + messages.link(bot.messenger, "https://goo.gl/Va817H", "Click here to open the chart")
 
@@ -201,26 +201,26 @@ def status(bot):
 
     return (
     "First you should check your node's status by running "
-    "<c>smartcash-cli smartnode status<c> on your node's VPS.\n\n"
+    "<c>stashcash-cli stashnode status<c> on your node's VPS.\n\n"
     "The status of your node can hint towards the problem of why your "
     "node isn't running. Here are the known states and their likely causes:\n\n"
     "<b>Node just started, not yet activated<b>\n\n"
     "Simply means your blocks aren't up-to-date yet with the current "
-    "blockcount. Run <c>smartcash-cli getinfo<c> on the VPS to see your current "
+    "blockcount. Run <c>stashcash-cli getinfo<c> on the VPS to see your current "
     "blockHeight at <c>blocks<c>, and compare it to the current blockHeight that "
     "you can see when you send me <cb>info<ca>\n\n"
     "<b>Broadcast IP doesn't match external IP<b>\n\n"
     "Most likely culprit is a duplicate IP/genkey/txhash in the "
-    "config file. Verify that all node entries in the <c>smartnode.conf<c> file "
+    "config file. Verify that all node entries in the <c>stashnode.conf<c> file "
     " have a unique IP/genkey/txhash+id. If your VPS has multiple IP's, adding "
-    "externalip=yourNodeIP in the <c>smartcash.conf<c> of a daemon will direct it"
+    "externalip=yourNodeIP in the <c>stashcash.conf<c> of a daemon will direct it"
     " towards the IP you specified.\n\n"
-    "<b>Not capable smartnode: invalid protocol version<b>\n\n"
+    "<b>Not capable stashnode: invalid protocol version<b>\n\n"
     "Your node is not up-to-date on the latest version. Compare the version you"
-    " see when you run <c>smartcash-cli getinfo<c> at <c>version<c> to the latest version at ") +\
-    messages.link(bot.messenger, "https://smartcash.cc/wallets/", "SmartCash wallets") +\
+    " see when you run <c>stashcash-cli getinfo<c> at <c>version<c> to the latest version at ") +\
+    messages.link(bot.messenger, "https://stashcash.cc/wallets/", "StashCash wallets") +\
     (" for more info on the current version.\n\n"
-    "<b>Not capable smartnode: smartnode not in smartnodelist<b>\n\n"
+    "<b>Not capable stashnode: stashnode not in stashnodelist<b>\n\n"
     "The node hasn't yet connected to the network. This can have "
     "several reasons; your desktop-wallet is outdated, your node is "
     "outdated or you didn't yet issue a <c>Start alias<c> command while the "
@@ -237,7 +237,7 @@ faqs = {
              initial ),
      'rewards' : FAQ("What payouts can i expect from my nodes?",
               rewards ),
-    'status' : FAQ(("What should be done when a SmartNode is <b>not<b> successfully started?"),
+    'status' : FAQ(("What should be done when a StashNode is <b>not<b> successfully started?"),
              status )
 }